Added a IT to ensure bookmark work as expected. Refactored the previous ConnectionIT (now DirectDriverIT)
